+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased
+----------
+- Allow for game pausing even while clearing lines
+
+
 0.2.0
 -----
 - Updated `winit` dependency to `0.30`